project-atlas/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
plotly==5.18.0
python-dateutil==2.8.2

